refactor: Move services out
