+++ src/azure/constants.rs
