+++ src/azure/mod.rs
@@ -1,0 +1,7 @@
+//! Azure Storage SharedKey support
+//!
+//! Use [`azure::storage::Signer`][crate::azure::storage::Signer]
+pub mod storage;
+
+mod constants;
+mod loader;
