+++ src/huaweicloud/constants.rs
