+++ src/huaweicloud/mod.rs
@@ -1,7 +1,7 @@
 //! Huawei Cloud signers
 //! Currently only support Object Storage Service (OBS) singer.
 //!
-//! Use [`services::huaweicloud::obs::Signer`][crate::services::huaweicloud::obs::Signer]
+//! Use [`huaweicloud::obs::Signer`][crate::huaweicloud::obs::Signer]
 pub mod obs;
 
 mod constants;
