--- src/services/azure/mod.rs
@@ -1,7 +1,0 @@
-//! Azure Storage SharedKey support
-//!
-//! Use [`services::azure::storage::Signer`][crate::services::azure::storage::Signer]
-pub mod storage;
-
-mod constants;
-mod loader;
