--- src/services/mod.rs
@@ -1,7 +1,0 @@
-//! Service implement of signers.
-
-pub mod aliyun;
-pub mod aws;
-pub mod azure;
-pub mod google;
-pub mod huaweicloud;
