+++ tests/aliyun/oss.rs
@@ -2,9 +2,9 @@ use anyhow::Result;
 use http::{Request, StatusCode};
 use log::{debug, warn};
 use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
+use reqsign::aliyun::loader;
+use reqsign::aliyun::oss::{Builder, Signer};
 use reqsign::credential::CredentialLoad;
-use reqsign::services::aliyun::loader;
-use reqsign::services::aliyun::oss::{Builder, Signer};
 use reqwest::blocking::Client;
 use serde::Deserialize;
 use std::env;
