+++ tests/aws/v4.rs
@@ -2,8 +2,8 @@ use anyhow::Result;
 use http::{Request, StatusCode};
 use log::{debug, warn};
 use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
-use reqsign::services::aws::loader;
-use reqsign::services::aws::v4::Signer;
+use reqsign::aws::loader;
+use reqsign::aws::v4::Signer;
 use reqwest::blocking::Client;
 use serde::Deserialize;
 use std::str::FromStr;
